🏷️ Fixed Tags & writing permissions
During development, an issue was found in adding multiple tags, where only one of two or more tags actually gets added to the note. On top of that, due to a reset of both the iOS and Android folders, the writing permissions had been removed; they are now back to normal.